Change addButton action listener to not spawn a new thread
Swing is not thread safe, and the thread that addButton.actionListener was kicking off was ultimately making modifcations to swing.

There are two proposed solutions. One is to wrap the Swing calls in some thread protection with `SwingUtilities.invokeLater()` to force it to run in the AWT thread.

The other solution is to remove the threading entirely. I'm almost 100% positive we don't need threads here, so I'm going to just remove the thread